+++ test/e2e/buckets.sh
@@ -122,6 +122,7 @@ EOF
 
 bucket_bot() { cat <<'EOF'
 TestGameServer_MinecraftBotConnects
+TestGameServer_TerrariaBotConnects
 EOF
 }